Updated fifo thread to gracefully recapture pipe     Note behavior of always filling fifo buffer

+++ stegasaur.py
@@ -1,5 +1,6 @@
 import argparse
 import os
+import select
 import threading
 import time
 
@@ -15,33 +16,42 @@ def getPipename(_pipename, _appendPid):
     rtn += ".pipe"
     return rtn
 
-def getStream(args):
-    if args.http:
-        return stegahttp.WebDataStream(args.http)
+def getStream(_args):
+    if _args.http:
+        return stegahttp.WebDataStream(_args.http)
 
-def refillPipe(_fifoPath, _stream, _chunkSize=1024):
-    # Open the FIFO for writing
-    with open(_fifoPath, 'wb') as fifo:
+def refillPipe(_fifoPath, _stream, _chunkSize=4096):
+    try:
+        fifo = open(_fifoPath, 'wb')
         while True:
-            if os.stat(_fifoPath).st_size == 0:
+            select.select([],[fifo],[])[1]
+            try:
                 fifo.write(_stream.read(_chunkSize))
                 fifo.flush()
+            except(BrokenPipeError):
+                fifo = open(_fifoPath, 'wb')
+    finally:
+        try:
+            os.remove(_fifoPath)
+        except:
+            pass
 
 def main(args):
     pipename = getPipename(args.pipename, args.appendPid)
-    os.remove(pipename)
+    if os.path.exists(pipename):
+        os.remove(pipename)
     os.mkfifo(pipename)
-    stream = getStream(args.generationType)
+    stream = getStream(args)
 
     refillThread = threading.Thread(target=refillPipe, args=(pipename, stream, args.chunk_size))
     refillThread.start()
 
 parser = argparse.ArgumentParser(description="Named pipe data generator")
-generationGroup = parser.add_mutually_exclusive_group(required=True)
-generationGroup.add_argument("--http", nargs="+")
 parser.add_argument("-P", "--no-append-pid", action="store_false",dest="appendPid", default=True, help="Append process ID to the pipe name (default: True)")
 parser.add_argument("-s", "--chunk-size", type=int, default=4096, help="Refill chunk size in bytes (default: 4096)")
 parser.add_argument("pipename", nargs="?", help="Name of the named pipe")
+generationGroup = parser.add_mutually_exclusive_group(required=True)
+generationGroup.add_argument("--http", nargs="+")
 
 if __name__ == "__main__":
     main(parser.parse_args())
